feat: device authentication (#42)
Implements device auth flow. As far as the state changes and interface is concerned, device auth flow is the same as the current browser challenge flow. The main advantage of device flow is that it avoids race conditions with the regular browser token because it is treated differently in the backend. Device token feature is not currently deployed on juliahub.com so this PR needs to be tested on development instances.

### Working of this PR:
Please see the changes in auth-flow.md file.
  
### Authenticating with device tokens:
The interface does not change for the user. `PkgAuthentication.authenticate()` can be called and it will seamlessly decide on what authentication mechanism to use.